+++ src/lang/es.ts
@@ -1,0 +1,21 @@
+import {Lang} from "./Lang";
+// {code: 'es', display: 'Spanish', flag: ''},
+
+export const es: Lang = {
+    code: 'es',
+    display: 'Spanish',
+    flag: '🇪🇸',
+    numbers: [
+			'cero',
+			'uno',
+			'dos',
+			'tres',
+			'cuatro',
+			'cinco',
+			'seis',
+			'siete',
+			'ocho',
+			'nueve',
+			'diez',
+		],
+}